:art: Rename resource description files
