+++ packages/nodes-base/nodes/Raindrop/descriptions/index.ts
@@ -1,4 +1,4 @@
-export * from './Collection';
-export * from './Raindrop';
-export * from './Tag';
-export * from './User';
+export * from './CollectionDescription';
+export * from './RaindropDescription';
+export * from './TagDescription';
+export * from './UserDescription';
